fixed py_model tests
